+++ 7run/UnixV6/OS.sys.cpp
@@ -162,7 +162,7 @@ int OS::syscall(int *result, int n, int arg0, uint8_t *args) {
 }
 
 int OS::v6_seek(int fd, off_t o, int w) { // 19
-    if (trace) fprintf(stderr, "<lseek(%d, %ld, %d)", fd, o, w);
+    if (trace) fprintf(stderr, "<lseek(%d, %ld, %d)", fd, long(o), w);
     FileBase *f = file(fd);
     off_t result = -1;
     switch (w) {
@@ -188,7 +188,7 @@ int OS::v6_seek(int fd, off_t o, int w) { // 19
             errno = EINVAL;
             break;
     }
-    if (trace) fprintf(stderr, " => %ld>\n", result);
+    if (trace) fprintf(stderr, " => %ld>\n", long(result));
     return result;
 }
